Move AcyclicOrientations inside orientations
